+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.4.7] 2026-02-01
+### Modified
+- Disk cache functions now detect saving backend when custom load/dump functions are provided.
+
+### Fixed
+- `Callable` in `isinstance_generic` check.
+
 ## [0.4.6] 2026-01-09
 ### Added
 - `cache_fname_fmt` can now be a custom callable formatter.